pyproject.toml
[project]
name = "mcplearning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.16",
    "fastmcp>=2.14.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
